requirements.txt
customtkinter>=5.2.0
Pillow>=9.0.0
tkcalendar>=1.6.1
